requirements.txt
Pillow
flask
beautifulsoup4
requests
